// examples/client/models/guid-file-dto.ts
/* tslint:disable */
/* eslint-disable */



export interface GuidFileDto {
    'id'?: string;
    'createTime'?: string;
    'createUser'?: string | null;
    'name'?: string | null;
    'key'?: string | null;
    'url'?: string | null;
    'size'?: number;
    'owner'?: string | null;
    'parentId'?: string | null;
    'isFolder'?: boolean;
}